+++ qddate/patterns/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-__version__ = '0.1.1'
+__version__ = "0.1.1"
 __author__ = "Ivan Begtin (ivan@begtin.tech)"
 __license__ = "BSD"
 
@@ -11,8 +11,19 @@ from .fr import PATTERNS_FR
 from .it import PATTERNS_IT
 from .pt import PATTERNS_PT
 from .ru import PATTERNS_RU
-from .base import PATTERNS_EN, BASE_TIME_PATTERNS
+from .base import PATTERNS_EN, BASE_TIME_PATTERNS, INTEGER_LIKE_PATTERNS
 
-ALL_PATTERNS = PATTERNS_EN + PATTERNS_BG + PATTERNS_CZ + PATTERNS_DE + PATTERNS_ES + PATTERNS_FR + PATTERNS_IT + PATTERNS_PT + PATTERNS_RU
+ALL_PATTERNS = (
+    PATTERNS_EN
+    + INTEGER_LIKE_PATTERNS
+    + PATTERNS_BG
+    + PATTERNS_CZ
+    + PATTERNS_DE
+    + PATTERNS_ES
+    + PATTERNS_FR
+    + PATTERNS_IT
+    + PATTERNS_PT
+    + PATTERNS_RU
+)
 
-SUPPORTED_LANGUAGES = ['bg', 'cz', 'de', 'en', 'fr', 'it', 'pt', 'ru']
+SUPPORTED_LANGUAGES = ["bg", "cz", "de", "en", "fr", "it", "pt", "ru"]